Refactor from chain of responsibility to observer, since each handler is not responsible any more on deciding to move to the next handler.

+++ src/main/java/br/com/granzoto/media_compressor/cloud_client_observer_handler/PdfCompressorHandler.java
@@ -1,5 +1,8 @@
-package br.com.granzoto.media_compressor.workflow;
+package br.com.granzoto.media_compressor.cloud_client_observer_handler;
 
+import br.com.granzoto.media_compressor.cloud_client.CloudClient;
+import br.com.granzoto.media_compressor.cloud_client.CloudClientHandler;
+import br.com.granzoto.media_compressor.cloud_client.CloudClientItemObserver;
 import br.com.granzoto.media_compressor.model.CompressionFile;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -7,26 +10,22 @@ import org.slf4j.LoggerFactory;
 import java.io.File;
 import java.util.Arrays;
 
-public class PdfCompressorHandler extends AbstractCloudClientHandler {
+public class PdfCompressorHandler implements CloudClientHandler, CloudClientItemObserver {
     private static final Logger LOGGER = LoggerFactory.getLogger(PdfCompressorHandler.class.getName());
     public static final String PDF_MIME_TYPE = "application/pdf";
 
     @Override
     public void handleItem(CompressionFile compressionFile) {
-        if (PDF_MIME_TYPE.equals(compressionFile.mimeType()) && !compressionFile.compressedFile().exists()){
+        if (PDF_MIME_TYPE.equals(compressionFile.mimeType()) && !compressionFile.compressedFile().exists()) {
             boolean executeCompression = this.executeCompression(compressionFile.originalFile(), compressionFile.compressedFile());
-            if (executeCompression) {
-                this.nextItemHandler(compressionFile);
-            } else if (compressionFile.compressedFile().exists()){
-                if (!compressionFile.compressedFile().delete()){
+            if (!executeCompression && compressionFile.compressedFile().exists()) {
+                if (compressionFile.compressedFile().delete()) {
+                    LOGGER.info("Invalid compressed file successfully deleted {}", compressionFile.compressedFile().getAbsolutePath());
+                } else {
                     LOGGER.warn("Fail deleting invalid compressed file {}. Will try to delete on JVM exit", compressionFile.compressedFile().getAbsolutePath());
                     compressionFile.compressedFile().deleteOnExit();
-                } else {
-                    LOGGER.info("Invalid compressed file successfully deleted {}", compressionFile.compressedFile().getAbsolutePath());
                 }
             }
-        } else {
-            this.nextItemHandler(compressionFile);
         }
     }
 
@@ -57,6 +56,11 @@ public class PdfCompressorHandler extends AbstractCloudClientHandler {
             LOGGER.error("PDF compression failed: {}. File: {}", e.getMessage(), inputFile.getAbsolutePath());
             return false;
         }
+    }
+
+    @Override
+    public void registerObserver(CloudClient cloudClient) {
+        cloudClient.addItemObserver(this);
     }
 }
