+++ src/main/java/br/com/granzoto/media_compressor/cloud_client_observer_handler/VideoCompressorHandler.java
@@ -1,41 +1,32 @@
-package br.com.granzoto.media_compressor.workflow;
+package br.com.granzoto.media_compressor.cloud_client_observer_handler;
 
 import br.com.granzoto.media_compressor.cloud_client.CloudClient;
+import br.com.granzoto.media_compressor.cloud_client.CloudClientHandler;
+import br.com.granzoto.media_compressor.cloud_client.CloudClientItemObserver;
 import br.com.granzoto.media_compressor.model.CompressionFile;
-
-import java.io.File;
-import java.util.Arrays;
-import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class VideoCompressorHandler extends AbstractCloudClientHandler {
+import java.io.File;
+import java.util.Arrays;
+
+public class VideoCompressorHandler implements CloudClientHandler, CloudClientItemObserver {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(VideoCompressorHandler.class.getName());
     public static final String VIDEO_MIME_TYPE = "video";
 
     @Override
-    public void handleStart(CloudClient cloudClient) {
-        super.handleStart(cloudClient);
-        this.nextStartHandler(cloudClient);
-    }
-
-    @Override
     public void handleItem(CompressionFile compressionFile) {
-        if (VIDEO_MIME_TYPE.equals(compressionFile.mimeSuperType()) && !compressionFile.compressedFile().exists()){
+        if (VIDEO_MIME_TYPE.equals(compressionFile.mimeSuperType()) && !compressionFile.compressedFile().exists()) {
             boolean executeCompression = this.executeCompression(compressionFile.originalFile(), compressionFile.compressedFile());
-            if (executeCompression) {
-                this.nextItemHandler(compressionFile);
-            } else if (compressionFile.compressedFile().exists()){
-                if (!compressionFile.compressedFile().delete()){
+            if (!executeCompression && compressionFile.compressedFile().exists()) {
+                if (compressionFile.compressedFile().delete()) {
+                    LOGGER.info("Invalid compressed file successfully deleted {}", compressionFile.compressedFile().getAbsolutePath());
+                } else {
                     LOGGER.warn("Fail deleting invalid compressed file {}. Will try to delete on JVM exit", compressionFile.compressedFile().getAbsolutePath());
                     compressionFile.compressedFile().deleteOnExit();
-                } else {
-                    LOGGER.info("Invalid compressed file successfully deleted {}", compressionFile.compressedFile().getAbsolutePath());
                 }
             }
-        } else {
-            this.nextItemHandler(compressionFile);
         }
     }
 
@@ -68,9 +59,7 @@ public class VideoCompressorHandler extends AbstractCloudClientHandler {
     }
 
     @Override
-    public void handleEnd(List<CompressionFile> files) {
-        super.handleEnd(files);
-        this.nextEndHandler(files);
+    public void registerObserver(CloudClient cloudClient) {
+        cloudClient.addItemObserver(this);
     }
-
 }
